+++ src/main/java/lotto/domain/purchase/Cash.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.domain.purchase;
 
 import lotto.domain.statistics.RateOfReturn;
