+++ src/test/java/lotto/domain/purchase/AmountTest.java
@@ -1,5 +1,6 @@
-package lotto.domain;
+package lotto.domain.purchase;
 
+import lotto.domain.purchase.Amount;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
