+++ src/test/java/lotto/domain/purchase/CashTest.java
@@ -1,5 +1,7 @@
-package lotto.domain;
+package lotto.domain.purchase;
 
+import lotto.domain.purchase.Amount;
+import lotto.domain.purchase.Cash;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
