+++ src/test/java/lotto/domain/statistics/WinningStatisticsTest.java
@@ -1,6 +1,6 @@
 package lotto.domain.statistics;
 
-import lotto.domain.Cash;
+import lotto.domain.purchase.Cash;
 import lotto.domain.rank.Rank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
